Ad Mob Added 1 Banner Ad 1 Interstitial Ad

+++ app/src/main/java/com/example/sntviaocr/ApplicationLayer/Notes/MainNotesDashboard.java
@@ -10,6 +10,7 @@ import androidx.lifecycle.ViewModelProvider;
 import androidx.recyclerview.widget.LinearLayoutManager;
 
 import android.Manifest;
+import android.annotation.SuppressLint;
 import android.content.Intent;
 import android.content.pm.PackageManager;
 import android.graphics.Bitmap;
@@ -33,6 +34,16 @@ import com.example.sntviaocr.BussinessLayer.ViewModels.NoteViewModel;
 import com.example.sntviaocr.DataLayer.Note.Note;
 import com.example.sntviaocr.R;
 import com.example.sntviaocr.databinding.ActivityMainNotesDashboardBinding;
+import com.google.android.gms.ads.AdError;
+import com.google.android.gms.ads.AdListener;
+import com.google.android.gms.ads.AdRequest;
+import com.google.android.gms.ads.FullScreenContentCallback;
+import com.google.android.gms.ads.LoadAdError;
+import com.google.android.gms.ads.MobileAds;
+import com.google.android.gms.ads.initialization.InitializationStatus;
+import com.google.android.gms.ads.initialization.OnInitializationCompleteListener;
+import com.google.android.gms.ads.interstitial.InterstitialAd;
+import com.google.android.gms.ads.interstitial.InterstitialAdLoadCallback;
 import com.google.android.gms.auth.api.signin.GoogleSignInAccount;
 import com.google.android.material.navigation.NavigationView;
 import com.google.mlkit.vision.common.InputImage;
@@ -43,15 +54,19 @@ import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.util.ArrayList;
+import java.util.Objects;
 
 public class MainNotesDashboard extends BasicActivityClass implements View.OnClickListener, MenuItem.OnMenuItemClickListener, NavigationView.OnNavigationItemSelectedListener, ICallback {
 
+    private static final String TAG = "MainNotesDashboard";
     public static final int NEW_NOTE_ACTIVITY_REQUEST_CODE = 2;
     private static final int MY_PERMISSIONS_REQUEST_READ_MEDIA = 101;
     private final int EDIT_NOTE_REQUEST = 102;
     private ActivityMainNotesDashboardBinding thisActivity;
     final int thumbSize = 128;
     private Observer<Integer> o;
+    private InterstitialAd mInterstitialAd;
+    private AdRequest adRequest;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -60,6 +75,8 @@ public class MainNotesDashboard extends BasicActivityClass implements View.OnCli
         thisActivity = ActivityMainNotesDashboardBinding.inflate(getLayoutInflater());
         setContentView(thisActivity.getRoot());
 
+        thisActivity.topAppBar.setTitle("Notes : " + display_name);
+
         recyclerView = findViewById(R.id.notes_list_recyclerview);
         noteListAdapter = new NoteListAdapter(new NoteListAdapter.NoteDiff());
         recyclerView.setAdapter(noteListAdapter);
@@ -67,6 +84,74 @@ public class MainNotesDashboard extends BasicActivityClass implements View.OnCli
 
         mNoteViewModel = new ViewModelProvider(this).get(NoteViewModel.class);
 
+        MobileAds.initialize(this, initializationStatus -> Log.d(TAG, "onInitializationComplete: " + initializationStatus.getAdapterStatusMap().toString()));
+
+        thisActivity.mainDashBoardAdView.setAdListener(new AdListener(){
+            @Override
+            public void onAdLoaded() {
+                super.onAdLoaded();
+                Log.d(TAG, "onAdLoaded: called");
+            }
+
+            @Override
+            public void onAdFailedToLoad(LoadAdError loadAdError) {
+                super.onAdFailedToLoad(loadAdError);
+                Log.d(TAG, "onAdFailedToLoad: called " + loadAdError.getMessage());
+            }
+
+            @Override
+            public void onAdOpened() {
+                super.onAdOpened();
+                Log.d(TAG, "onAdOpened: called");
+            }
+
+            @Override
+            public void onAdClosed() {
+                super.onAdClosed();
+                Log.d(TAG, "onAdClosed: called");
+            }
+        });
+        adRequest = new AdRequest.Builder().build();
+
+        thisActivity.mainDashBoardAdView.loadAd(adRequest);
+
+        InterstitialAd.load(this, "ca-app-pub-8387945083932065/6681386823", new AdRequest.Builder().build(),new InterstitialAdLoadCallback(){
+            @Override
+            public void onAdLoaded(@NonNull InterstitialAd interstitialAd) {
+                super.onAdLoaded(interstitialAd);
+                mInterstitialAd = interstitialAd;
+                mInterstitialAd.setFullScreenContentCallback(new FullScreenContentCallback(){
+                    @Override
+                    public void onAdDismissedFullScreenContent() {
+                        // Called when fullscreen content is dismissed.
+                        Log.d(TAG, "The ad was dismissed.");
+                    }
+
+                    @Override
+                    public void onAdFailedToShowFullScreenContent(AdError adError) {
+                        // Called when fullscreen content failed to show.
+                        Log.d(TAG, "The ad failed to show.");
+                    }
+
+                    @Override
+                    public void onAdShowedFullScreenContent() {
+                        // Called when fullscreen content is shown.
+                        // Make sure to set your reference to null so you don't
+                        // show it a second time.
+                        mInterstitialAd = null;
+                        Log.d(TAG, "The ad was shown.");
+                    }
+                });
+                Log.d(TAG, "InterstitialAd onAdLoaded: called" + interstitialAd.getResponseInfo());
+            }
+
+            @Override
+            public void onAdFailedToLoad(@NonNull LoadAdError loadAdError) {
+                super.onAdFailedToLoad(loadAdError);
+                mInterstitialAd = null;
+                Log.d(TAG, "InterstitialAd onAdFailedToLoad: " + loadAdError.getMessage());
+            }
+        });
 
         mNoteViewModel.getAllNotes()
                         .observe(this, notes -> {
@@ -75,18 +160,15 @@ public class MainNotesDashboard extends BasicActivityClass implements View.OnCli
                         });
 
 
-        Observer<Integer> o = new Observer<Integer>() {
-            @Override
-            public void onChanged(Integer note_id) {
-                if(observerCounter > 0){
-                    Log.i("MUTEABLELIVEDATA", "INTEGER CHANGED : " + note_id);
-                    Intent i = new Intent(MainNotesDashboard.this, EditAnySaveNoteActivity.class);
-                    i.putExtra("NoteId", BasicActivityClass.NoteIdClicked.getValue());
-                    i.putExtra("ClickedNote", ClickedNote);
-                    startActivityForResult(i,EDIT_NOTE_REQUEST);
-                }
-                observerCounter++;
+        Observer<Integer> o = note_id -> {
+            if(observerCounter > 0){
+                Log.i("MUTEABLELIVEDATA", "INTEGER CHANGED : " + note_id);
+                Intent i = new Intent(MainNotesDashboard.this, EditAnySaveNoteActivity.class);
+                i.putExtra("NoteId", BasicActivityClass.NoteIdClicked.getValue());
+                i.putExtra("ClickedNote", ClickedNote);
+                startActivityForResult(i,EDIT_NOTE_REQUEST);
             }
+            observerCounter++;
         };
 
         NoteIdClicked.observe(this, o);
@@ -98,12 +180,7 @@ public class MainNotesDashboard extends BasicActivityClass implements View.OnCli
         thisActivity.logoutButton.setOnClickListener(this);
         thisActivity.takePhotoFabButton.setOnClickListener(this);
 //        thisActivity.bottomAppBar.getMenu().findItem(R.id.rescan_photo_button).setOnMenuItemClickListener(this);
-        thisActivity.bottomAppBar.setNavigationOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                thisActivity.drawerLayout.openDrawer(Gravity.LEFT);
-            }
-        });
+        thisActivity.bottomAppBar.setNavigationOnClickListener(v -> thisActivity.drawerLayout.openDrawer(GravityCompat.START));
 
     }
 
@@ -111,7 +188,7 @@ public class MainNotesDashboard extends BasicActivityClass implements View.OnCli
     @Override
     public void onBackPressed() {
         super.onBackPressed();
-        if(NoteViewHolder.speaking == true)
+        if(NoteViewHolder.speaking)
         {
             BasicActivityClass.stopSpeaking();
         }
@@ -122,7 +199,7 @@ public class MainNotesDashboard extends BasicActivityClass implements View.OnCli
     {
 
         // declaration and initialise String Array
-        String str[] = new String[arr.size()];
+        String[] str = new String[arr.size()];
 
         // ArrayList to Array Conversion
         for (int j = 0; j < arr.size(); j++) {
@@ -146,8 +223,8 @@ public class MainNotesDashboard extends BasicActivityClass implements View.OnCli
     @Override
     public void UpdateUI(GoogleSignInAccount account) {
         if(account != null) {
-        }
-        else{
+            Log.d(TAG, "UpdateUI: " + "Account not null");
+        } else{
             Intent mainToSignInIntent = new Intent(MainNotesDashboard.this, SplashScreen.class);
             startActivity(mainToSignInIntent);
             finish();
@@ -155,16 +232,11 @@ public class MainNotesDashboard extends BasicActivityClass implements View.OnCli
     }
 
     @Override
-    public void onRequestPermissionsResult(int requestCode, String permissions[], int[] grantResults) {
-        switch (requestCode) {
-            case MY_PERMISSIONS_REQUEST_READ_MEDIA:
-                if ((grantResults.length > 0) && (grantResults[0] == PackageManager.PERMISSION_GRANTED)) {
-                    cameraHandler.deleteLastPhotoTaken(getApplicationContext());
-                }
-                break;
-
-            default:
-                break;
+    public void onRequestPermissionsResult(int requestCode, @NonNull String[] permissions, @NonNull int[] grantResults) {
+        if (requestCode == MY_PERMISSIONS_REQUEST_READ_MEDIA) {
+            if ((grantResults.length > 0) && (grantResults[0] == PackageManager.PERMISSION_GRANTED)) {
+                cameraHandler.deleteLastPhotoTaken(getApplicationContext());
+            }
         }
     }
 
@@ -185,8 +257,6 @@ public class MainNotesDashboard extends BasicActivityClass implements View.OnCli
                     out.flush();
                     out.close();
                 }
-            } catch (FileNotFoundException e) {
-                e.printStackTrace();
             } catch (IOException e) {
                 e.printStackTrace();
             }
@@ -199,7 +269,7 @@ public class MainNotesDashboard extends BasicActivityClass implements View.OnCli
 
         }
         else if(requestCode == NEW_NOTE_ACTIVITY_REQUEST_CODE) {
-            int button_clicked = data.getIntExtra(EditViewOfNote.ACTIVITY_BUTTON_CLICKED_REPLY , 0);
+            int button_clicked = Objects.requireNonNull(data).getIntExtra(EditViewOfNote.ACTIVITY_BUTTON_CLICKED_REPLY , 0);
             if(button_clicked == R.id.edit_note_view_save_button){
                 String s = data.getStringExtra(EditViewOfNote.ACTIVITY_EXTRACTED_TEXT_TITLE_REPLY);
                 Note note = new Note(
@@ -228,16 +298,11 @@ public class MainNotesDashboard extends BasicActivityClass implements View.OnCli
         }else if(requestCode == EDIT_NOTE_REQUEST){
             //TODO SAVE EDITED NOTE
             try {
-                Note n = (Note) data.getSerializableExtra("EditedNote");
+                Note n = (Note) Objects.requireNonNull(data).getSerializableExtra("EditedNote");
                 if(n != null){
                     ClickedNote = n;
 //                    mNoteViewModel.updateNote(ClickedNote.NoteId, ClickedNote.Title, ClickedNote.Content);
-                    AsyncTask.execute(new Runnable() {
-                        @Override
-                        public void run() {
-                            mNoteViewModel.updateNote(ClickedNote.NoteId, ClickedNote.Title, ClickedNote.Content);
-                        }
-                    });
+                    AsyncTask.execute(() -> mNoteViewModel.updateNote(ClickedNote.NoteId, ClickedNote.Title, ClickedNote.Content));
                 }
             }catch (Exception ex){
                 ex.printStackTrace();
@@ -258,18 +323,13 @@ public class MainNotesDashboard extends BasicActivityClass implements View.OnCli
 
     @Override
     public boolean onMenuItemClick(MenuItem item) {
-        switch (item.getItemId())
-        {
-//            case R.id.rescan_photo_button:
-//                try{
-//                    processImage(cameraHandler.getImage(this));
-//                }catch(NullPointerException ex){
-//                    ShowMessage("Error", ex.getMessage(),this);
-//                }
-//                return true;
-
-            default:break;
-        }
+        //            case R.id.rescan_photo_button:
+        //                try{
+        //                    processImage(cameraHandler.getImage(this));
+        //                }catch(NullPointerException ex){
+        //                    ShowMessage("Error", ex.getMessage(),this);
+        //                }
+        //                return true;
         return false;
     }
 
@@ -289,6 +349,10 @@ public class MainNotesDashboard extends BasicActivityClass implements View.OnCli
                 //TODO Implement Settings
 //                break;
 
+            case R.id.menu_item_help:
+                startActivity(new Intent(this, HelpActivity.class));
+                break;
+
             case R.id.menu_item_logout:
                 signOut();
 //                NoteIdClicked.removeObserver(o);
@@ -301,6 +365,7 @@ public class MainNotesDashboard extends BasicActivityClass implements View.OnCli
         return true;
     }
 
+    @SuppressLint("NonConstantResourceId")
     @Override
     public void onClick(View v) {
         switch (v.getId())
@@ -314,7 +379,12 @@ public class MainNotesDashboard extends BasicActivityClass implements View.OnCli
             case R.id.take_photo_fab_button:
 //                Snackbar.make(findViewById(android.R.id.content),"SNACKBAR DEMO",Snackbar.LENGTH_LONG).show();
                 mlKit.setIc(this);
-                startActivityForResult(cameraHandler.GetPictureIntent(activity), CameraHandler.getCameraHandlerRequestCode());
+                if (mInterstitialAd != null) {
+                    mInterstitialAd.show(MainNotesDashboard.this);
+                } else {
+                    Log.d(TAG, "The interstitial ad wasn't ready yet.");
+                    startActivityForResult(cameraHandler.GetPictureIntent(activity), CameraHandler.getCameraHandlerRequestCode());
+                }
                 break;
 
             default:break;
